+++ packages/kokoas-client/src/pages/projContracts/hooks/useContractProcess.ts
@@ -19,7 +19,6 @@ export const useContractProcess = () => {
     sendContract,
     {
       onMutate: () => {
-        console.log('Starting!');
         setBackdropState({ open: true });
       },
