+++ packages/kokoas-client/src/pages/projContracts/hooks/useHandleProjEstimate.ts
@@ -3,11 +3,11 @@ import { useFormikContext } from 'formik';
 import { useState } from 'react';
 import { calculateEstimate } from '../../../api/others/calculateEstimate';
 import {  TypeOfForm } from '../form';
-import { normalizedData } from '../api/getFormDataId';
 import useDeepCompareEffect from 'use-deep-compare-effect';
 import isEmpty from 'lodash/isEmpty';
 import { useEstimateById } from '../../../hooksQuery/useEstimateById';
 import { IProjestimates } from 'types';
+import { convertToForm } from '../api/convertToForm';
 
 
 /**
@@ -42,7 +42,7 @@ export const useHandleProjEstimate = () => {
       const {
         newFormData,
         newCalculated,
-      } = normalizedData(selectedRecord, calculated);
+      } = convertToForm(selectedRecord, calculated);
 
       setValues((prev) => ({
         ...prev,
